+++ resources/views/auth/pwa-resync.blade.php
@@ -1,0 +1,5 @@
+@extends('layouts.app')
+
+@section('content')
+    <x-easy-auth::auth.pwa-resync-panel />
+@endsection